+++ src/TransferService.java
@@ -1,3 +1,5 @@
+import java.math.BigDecimal;
+
 public class TransferService {
     private final Notifier notifier;
 
@@ -9,7 +11,7 @@ public class TransferService {
         this.notifier = notifier;
     }
 
-    public void transfer(BankAccount from, BankAccount to, double amount) {
+    public void transfer(BankAccount from, BankAccount to, BigDecimal amount) {
         from.withdraw(amount);
         to.deposit(amount);
         notifier.send("Wyslano przelew");
